chore: lift common functionality
