+++ src/grpc/mod.rs
@@ -1,11 +1,18 @@
-use crate::{db, jwt::JwtVerifier, middleware::AuthLayer, Context};
-use std::{fs::read_to_string, net::SocketAddr, sync::Arc};
+use crate::{
+    db,
+    jwt::JwtVerifier,
+    middleware::AuthLayer,
+    proto::common::{ChartData as PbChartData, Rating as PbRating},
+    ratings::{get_snap_name, ChartData},
+    Context,
+};
+use futures::future::try_join_all;
+use std::{error::Error, fs::read_to_string, net::SocketAddr, sync::Arc};
 use tonic::{
     transport::{Identity, Server, ServerTlsConfig},
     Status,
 };
-use tracing::warn;
-
+use tracing::{error, warn};
 mod app;
 mod charts;
 mod user;
@@ -54,4 +61,45 @@ pub async fn run_server(ctx: Context) -> Result<(), Box<dyn std::error::Error>> 
         .await?;
 
     Ok(())
+}
+
+pub(crate) async fn populate_chart_data_with_names(
+    ctx: &Arc<Context>,
+    chart_data_vec: Vec<ChartData>,
+) -> Result<Vec<PbChartData>, Status> {
+    try_join_all(chart_data_vec.into_iter().map(|chart_data| async {
+        let snap_name = get_snap_name(
+            &chart_data.rating.snap_id,
+            &ctx.config.snapcraft_io_uri,
+            &ctx.http_client,
+        )
+        .await
+        .map_err(|e| {
+            let mut err = &e as &dyn Error;
+            let mut error_chain = format!("{err}");
+            while let Some(src) = err.source() {
+                error_chain.push_str(&format!("\nCaused by: {src}"));
+                err = src;
+            }
+            error!(error=%error_chain, "unable to fetch snap name");
+            Status::unknown("Internal server error")
+        })?;
+
+        Ok(PbChartData::from_chart_data_and_snap_name(
+            chart_data, snap_name,
+        ))
+    }))
+    .await
+}
+
+impl PbChartData {
+    pub fn from_chart_data_and_snap_name(chart_data: ChartData, snap_name: String) -> Self {
+        Self {
+            raw_rating: chart_data.raw_rating,
+            rating: Some(PbRating::from_rating_and_snap_name(
+                chart_data.rating,
+                snap_name,
+            )),
+        }
+    }
 }
